Change calendar logic
Make calendar a class on its own, not an
instance of Event. Only events are stored
in the db, a Calendar is a logic layer
above events. Events have a type (db changed),
so that 'Rassegna' events are then reunited
in a Calendar when read from the db.
This also solves the issue of publication
of calendar events proposed by non admins.
Notifications for that kind of events should
work, too.